Starting on our own recipes, getting my bearings.

+++ container/build.sh
@@ -8,25 +8,46 @@ ls -alh
 
 pushd yocto/poky
 
+ls -alh
+
 source oe-init-build-env build-wifx
+
+# time bitbake cameroon-ucla
+
+time bitbake wifx-base
 
 ls -alh
 
-time nice -n19 bitbake wifx-base
-
-find . -type d
+pwd
 
 IMAGES=`pwd`/tmp/deploy/images
 
 ROOTFS=`find . -name rootfs`
 
-pushd $ROOTFS
+for rpm in `find tmp/work/cortexa5hf-neon-poky-linux-gnueabi -name "*cameroon*.rpm"`; do
+	echo $rpm
+	cp $rpm $IMAGES
+done
 
-rm -f ${IMAGES}/rootfs.tar.bz2
+for dir in `find tmp/work/cortexa5hf-neon-poky-linux-gnueabi -name "*cameroon*" -type d`; do
+	find $dir
+done
 
-tar cpjf ${IMAGES}/rootfs.tar.bz2 .
+if [ -z "$ROOTFS" ]; then
+	echo "missing rootfs"
+else
+	rsync -ua ${ROOTFS}/ ${IMAGES}/rootfs/
 
-popd
+	pushd $ROOTFS
+
+	pwd
+
+	rm -f ${IMAGES}/rootfs.tar.bz2
+
+	tar cpjf ${IMAGES}/rootfs.tar.bz2 .
+
+	popd
+fi
 
 popd
